@@ -16,11 +16,10 @@
     "npm": ">=6 <7"
   },
   "scripts": {
-    "build": "webpack -p --mode production --https false",
-    "build:dev": "webpack --mode development --https false",
-    "build-dev": "webpack --mode development --https false && echo . && echo . && echo . && echo Please use 'build:dev' instead of 'build-dev'.",
+    "build": "webpack --mode production",
+    "build:dev": "webpack --mode development",
     "convert-to-single-host": "node convertToSingleHost.js",
-    "dev-server": "webpack-dev-server --mode development",
+    "dev-server": "webpack serve --mode development",
     "lint": "office-addin-lint check",
     "lint:fix": "office-addin-lint fix",
     "prettier": "office-addin-lint prettier",
@@ -34,7 +33,7 @@
     "watch": "webpack --mode development --watch"
   },
   "dependencies": {
-    "@fluentui/react": "^8.16.0",
+    "@fluentui/react": "^8.34.0",
     "core-js": "^3.9.1",
     "es6-promise": "^4.2.8",
     "react": "^16.8.2",
@@ -53,40 +52,41 @@
     "@types/react-dom": "^16.8.4",
     "@types/react-hot-loader": "^4.1.0",
     "@types/webpack": "^4.4.34",
-    "@types/webpack-dev-server": "^3.1.6",
+    "@types/webpack-dev-server": "^4.1.0",
+    "acorn": "^8.5.0",
     "babel-loader": "^8.2.2",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^6.4.1",
-    "css-loader": "^5.2.4",
+    "copy-webpack-plugin": "^9.0.1",
+    "css-loader": "^6.2.0",
     "eslint": "^7.20.0",
-    "eslint-plugin-office-addins": "^1.0.0",
+    "eslint-plugin-office-addins": "^1.0.1",
     "eslint-plugin-react": "^7.16.0",
-    "extract-text-webpack-plugin": "4.0.0-beta.0",
-    "file-loader": "^4.2.0",
+    "file-loader": "^6.2.0",
     "find-process": "^1.4.4",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^4.5.1",
+    "html-loader": "^2.1.2",
+    "html-webpack-plugin": "^5.3.2",
     "less": "^3.9.0",
-    "less-loader": "^5.0.0",
+    "less-loader": "^10.0.1",
+    "mini-css-extract-plugin": "^2.2.2",
     "mocha": "^8.2.1",
-    "office-addin-cli": "^1.2.2",
-    "office-addin-debugging": "^4.2.2",
-    "office-addin-dev-certs": "^1.6.2",
-    "office-addin-lint": "^1.3.2",
-    "office-addin-manifest": "^1.6.2",
+    "office-addin-cli": "^1.2.3",
+    "office-addin-debugging": "^4.2.3",
+    "office-addin-dev-certs": "^1.6.3",
+    "office-addin-lint": "^1.3.3",
+    "office-addin-manifest": "^1.6.3",
     "office-addin-prettier-config": "^1.1.0",
-    "office-addin-test-helpers": "^1.1.2",
-    "office-addin-test-server": "^1.1.2",
-    "postcss-loader": "^4.0.0",
-    "react-hot-loader": "^4.12.6",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^0.23.1",
-    "ts-loader": "^6.2.2",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.1.3",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.2"
+    "office-addin-test-helpers": "^1.1.3",
+    "office-addin-test-server": "^1.1.3",
+    "os-browserify": "^0.3.0",
+    "process": "^0.11.10",
+    "source-map-loader": "^3.0.0",
+    "style-loader": "^3.2.1",
+    "ts-loader": "^9.2.5",
+    "ts-node": "^10.2.0",
+    "typescript": "^4.3.5",
+    "webpack": "^5.50.0",
+    "webpack-cli": "^4.8.0",
+    "webpack-dev-server": "4.1.1"
   },
   "prettier": "office-addin-prettier-config"
 }
