@@ -1,11 +1,20 @@
+/* global require, process, console */
+
 const convertTest = process.argv[3] === "convert-test";
 const fs = require("fs");
 const host = process.argv[2];
 const hosts = ["excel", "onenote", "outlook", "powerpoint", "project", "word"];
 const path = require("path");
 const util = require("util");
-const testPackages = ["@types/mocha", "@types/node", "current-processes", "mocha", "office-addin-test-helpers",
-  "office-addin-test-server", "ts-node"];
+const testPackages = [
+  "@types/mocha",
+  "@types/node",
+  "current-processes",
+  "mocha",
+  "office-addin-test-helpers",
+  "office-addin-test-server",
+  "ts-node",
+];
 const readFileAsync = util.promisify(fs.readFile);
 const unlinkFileAsync = util.promisify(fs.unlink);
 const writeFileAsync = util.promisify(fs.writeFile);
@@ -25,7 +34,7 @@ async function modifyProjectForSingleHost(host) {
 }
 
 async function convertProjectToSingleHost(host) {
-  // copy host-specific manifest over manifest.xml  
+  // copy host-specific manifest over manifest.xml
   const manifestContent = await readFileAsync(`./manifest.${host}.xml`, "utf8");
   await writeFileAsync(`./manifest.xml`, manifestContent);
 
@@ -38,7 +47,10 @@ async function convertProjectToSingleHost(host) {
   if (convertTest && (host === "excel" || host === "word")) {
     // copy over host-specific taskpane test code to test-taskpane.ts
     const testTaskpaneContent = await readFileAsync(`./test/src/test.${host}.app.tsx`, "utf8");
-    const updatedTestTaskpaneContent = testTaskpaneContent.replace(`../../src/taskpane/components/${host}.App`, `../../src/taskpane/components/App`);
+    const updatedTestTaskpaneContent = testTaskpaneContent.replace(
+      `../../src/taskpane/components/${host}.App`, 
+      `../../src/taskpane/components/App`
+    );
     await writeFileAsync(`./test/src/test.app.tsx`, updatedTestTaskpaneContent);
 
     // update ui-test.ts to only run against specified host
@@ -52,8 +64,7 @@ async function convertProjectToSingleHost(host) {
         await unlinkFileAsync(`./test/src/test.${host}.app.tsx`);
       }
     });
-  }
-  else {
+  } else {
     deleteFolder(path.resolve(`./test`));
   }
 
@@ -83,16 +94,17 @@ async function updatePackageJsonForSingleHost(host) {
   delete content.engines;
 
   // update sideload and unload scripts to use selected host.
-  ["sideload", "unload"].forEach(key => {
+  ["sideload", "unload"].forEach((key) => {
     content.scripts[key] = content.scripts[`${key}:${host}`];
   });
 
   // remove scripts that are unrelated to the selected host
   Object.keys(content.scripts).forEach(function (key) {
-    if (key.startsWith("sideload:")
-      || key.startsWith("unload:")
-      || key === "convert-to-single-host"
-      || key === "start:desktop:outlook"
+    if (
+      key.startsWith("sideload:") ||
+      key.startsWith("unload:") ||
+      key === "convert-to-single-host" ||
+      key === "start:desktop:outlook"
     ) {
       delete content.scripts[key];
     }
@@ -109,7 +121,7 @@ async function updatePackageJsonForSingleHost(host) {
     // remove test-related packages
     Object.keys(content.devDependencies).forEach(function (key) {
       if (testPackages.includes(key)) {
-        delete content.devDependencies[key]
+        delete content.devDependencies[key];
       }
     });
   }
@@ -149,13 +161,12 @@ function getHostName(host) {
 function deleteFolder(folder) {
   try {
     if (fs.existsSync(folder)) {
-      fs.readdirSync(folder).forEach(function (file, index) {
+      fs.readdirSync(folder).forEach(function (file) {
         const curPath = `${folder}/${file}`;
 
         if (fs.lstatSync(curPath).isDirectory()) {
           deleteFolder(curPath);
-        }
-        else {
+        } else {
           fs.unlinkSync(curPath);
         }
       });
@@ -166,20 +177,19 @@ function deleteFolder(folder) {
   }
 }
 
-async function deleteSupportFiles()
-{
-    await unlinkFileAsync("CONTRIBUTING.md");
-    await unlinkFileAsync("LICENSE");
-    await unlinkFileAsync("README.md");
-    await unlinkFileAsync("./convertToSingleHost.js");
-    await unlinkFileAsync(".npmrc");
+async function deleteSupportFiles() {
+  await unlinkFileAsync("CONTRIBUTING.md");
+  await unlinkFileAsync("LICENSE");
+  await unlinkFileAsync("README.md");
+  await unlinkFileAsync("./convertToSingleHost.js");
+  await unlinkFileAsync(".npmrc");
 }
 
 /**
  * Modify the project so that it only supports a single host.
  * @param host The host to support.
  */
-modifyProjectForSingleHost(host).catch(err => {
+modifyProjectForSingleHost(host).catch((err) => {
   console.error(`Error: ${err instanceof Error ? err.message : err}`);
   process.exitCode = 1;
 });
