@@ -3,12 +3,17 @@
 const devCerts = require("office-addin-dev-certs");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
 const CopyWebpackPlugin = require("copy-webpack-plugin");
-const ExtractTextPlugin = require("extract-text-webpack-plugin");
+const ExtractCSSPlugin = require("mini-css-extract-plugin");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const webpack = require("webpack");
 
 const urlDev = "https://localhost:3000/";
 const urlProd = "https://www.contoso.com/"; // CHANGE THIS TO YOUR PRODUCTION DEPLOYMENT LOCATION
+
+async function getHttpsOptions() {
+  const httpsOptions = await devCerts.getHttpsServerOptions();
+  return { cacert: httpsOptions.ca, key: httpsOptions.key, cert: httpsOptions.cert };
+}
 
 module.exports = async (env, options) => {
   const dev = options.mode === "development";
@@ -21,6 +26,10 @@ module.exports = async (env, options) => {
       taskpane: ["react-hot-loader/patch", "./src/taskpane/index.tsx"],
       commands: "./src/commands/commands.ts",
     },
+    output: {
+      sourceMapFilename: "[name].js.map",
+      devtoolModuleFilenameTemplate: "webpack:///[resource-path]?[loaders]",
+    },
     resolve: {
       extensions: [".ts", ".tsx", ".html", ".js"],
     },
@@ -32,15 +41,17 @@ module.exports = async (env, options) => {
           exclude: /node_modules/,
         },
         {
-          test: /\.css$/,
-          use: ["style-loader", "css-loader"],
+          test: /\.html$/,
+          exclude: /node_modules/,
+          use: "html-loader",
         },
         {
-          test: /\.(png|jpg|jpeg|gif)$/,
-          loader: "file-loader",
-          options: {
-            name: "[path][name].[ext]",
-          },
+          test: /\.css$/,
+          use: [dev ? "style-loader" : ExtractCSSPlugin.loader, "css-loader"],
+        },
+        {
+          test: /\.(png|jpg|jpeg|gif|ico)$/,
+          type: "asset/resource",
         },
       ],
     },
@@ -54,7 +65,7 @@ module.exports = async (env, options) => {
           },
           {
             from: "manifest*.xml",
-            to: "[name]." + buildType + ".[ext]",
+            to: "[name]." + buildType + "[ext]",
             transform(content) {
               if (dev) {
                 return content;
@@ -65,7 +76,6 @@ module.exports = async (env, options) => {
           },
         ],
       }),
-      new ExtractTextPlugin("[name].[hash].css"),
       new HtmlWebpackPlugin({
         filename: "taskpane.html",
         template: "./src/taskpane/taskpane.html",
@@ -79,13 +89,13 @@ module.exports = async (env, options) => {
       new webpack.ProvidePlugin({
         Promise: ["es6-promise", "Promise"],
       }),
-    ],
+    ].concat(dev ? [] : [new ExtractCSSPlugin({ filename: "[name].[hash].css" })]),
     devServer: {
       hot: true,
       headers: {
         "Access-Control-Allow-Origin": "*",
       },
-      https: options.https !== undefined ? options.https : await devCerts.getHttpsServerOptions(),
+      https: env.WEBPACK_BUILD || options.https !== undefined ? options.https : await getHttpsOptions(),
       port: process.env.npm_package_config_dev_server_port || 3000,
     },
   };
